requirements.txt
azure-functions
openai
azure-cosmos
apify-client
requests
python-dotenv